pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlc-cloud-repos"
version = "0.1.0"
description = "A universal cloud-init querying and repository configuration tool for Rocky Linux CIQ Derivatives"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Joel Hanger", email = "jhanger@ciq.com" }]
dependencies = ["PyYAML", "ruamel.yaml"]

[project.scripts]
rlc-cloud-repos = "rlc_cloud_repos.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["rlc_cloud_repos", "rlc_cloud_repos.*"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
